wip: On-disk creation
