+++ relay-server/src/lib.rs
@@ -269,6 +269,7 @@ mod utils;
 pub use self::services::spooler::spool_utils;
 // Public just for benchmarks.
 pub use self::envelope::Envelope;
+pub use services::buffer::{EnvelopeStack, SqliteEnvelopeStack};
 
 #[cfg(test)]
 mod testutils;
